st - Rename integration_helper -> test_helper. Intend to use it for generator tests in addition to integration tests.
