+++ test/integration/db_fixtures_load_test.rb
@@ -1,4 +1,4 @@
-require 'integration_helper'
+require 'test_helper'
 require 'open3'
 
 class DbFixtureLoadTest < ActiveSupport::TestCase
